requirements.txt
heatshrink2>=0.12.0
pyyaml>=6.0.1
